Spun custom Swig stuff into own module
